requirements.txt
pandas==2.2.0
SQLAlchemy==2.0.25
mysql-connector-python==8.3.0
python-dotenv==1.0.1
